Module amazon_sp_listings_items::apis::listings_api

Enums

DeleteListingsItemError

struct for typed errors of method delete_listings_item

PatchListingsItemError

struct for typed errors of method patch_listings_item

PutListingsItemError

struct for typed errors of method put_listings_item

Functions

delete_listings_item

Delete a listings item for a selling partner. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 5 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

patch_listings_item

Partially update (patch) a listings item for a selling partner. Only top-level listings item attributes can be patched. Patching nested attributes is not supported. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 5 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

put_listings_item

Creates a new or fully-updates an existing listings item for a selling partner. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 5 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see Usage Plans and Rate Limits in the Selling Partner API.