Function amazon_sp_listings_items::apis::listings_api::put_listings_item

pub async fn put_listings_item(
    configuration: &Configuration,
    seller_id: &str,
    sku: &str,
    marketplace_ids: Vec<String>,
    body: ListingsItemPutRequest,
    issue_locale: Option<&str>
) -> Result<ListingsItemSubmissionResponse, Error>

Creates a new or fully-updates an existing listings item for a selling partner. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 5 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see Usage Plans and Rate Limits in the Selling Partner API.